+++ spec/nodeSpec.js
@@ -1,4 +1,4 @@
-describe('Node', function() {
+fdescribe('Node', function() {
     var GR = require('../src/graphTheory');
     var myNode;
 
@@ -10,24 +10,26 @@ describe('Node', function() {
             name: "NYC"
         });
     });
-
-
     describe('init', function() {
         it('initializes with a label', function() {
             expect(myNode.label).toEqual("NYC");
-        });
-        it('defaults label to a random integer above 32', function() {
-            var newV = new GR.Node();
-            expect(newV.label).toBeGreaterThan(32);
         });
         it('initialzies with a data object', function() {
             expect(myNode.data).toBeObject();
         });
     });
     describe('#setLabel', () => {
-        it('returns and sets label to a random integer above 32 ', function() {
+        it('returns and sets label  ', function() {
             let newNode = new GR.Node()
-            expect(newNode.label).toBeGreaterThan(32);
+            expect(newNode.label).toBeUndefined();
+        });
+    });
+    describe('setData(d)', () => {
+        it('sets the data attribute', function() {
+            myNode.setData({
+                name: 'nyc'
+            });
+            expect(myNode.data.name).toEqual('nyc');
         });
     });
 });